+++ src/clearskies/di/__init__.py
@@ -1,6 +1,7 @@
 import clearskies.di.inject as inject
 from clearskies.di.additional_config import AdditionalConfig
 from clearskies.di.additional_config_auto_import import AdditionalConfigAutoImport
+from clearskies.di.additional_mygrations_auto_import import AdditionalMygrationsAutoImport
 from clearskies.di.di import Di
 from clearskies.di.injectable import Injectable
 from clearskies.di.injectable_properties import InjectableProperties
@@ -8,6 +9,7 @@ from clearskies.di.injectable_properties import InjectableProperties
 __all__ = [
     "AdditionalConfig",
     "AdditionalConfigAutoImport",
+    "AdditionalMygrationsAutoImport",
     "Di",
     "Injectable",
     "InjectableProperties",